feat: Add MongoDB integration and refactor Razorpay client
