+++ .env.example
@@ -1,6 +1,11 @@
 # Server
 PORT=3000
 
+# MongoDB
+# Example for local MongoDB; or use your Atlas connection string
+# MONGO_URI=mongodb://localhost:27017/foodapp
+MONGO_URI=
+
 # Razorpay Keys (obtain from Razorpay Dashboard)
 RAZORPAY_KEY_ID=rzp_test_xxxxxxxxxxxxx
 RAZORPAY_KEY_SECRET=xxxxxxxxxxxxxxxxxxxx
